requirements.txt
django
django-filter
django-extensions
gunicorn
djangorestframework
urllib3
twilio
numpy
firebase_admin